refactor: interview 도메인 필드명·타입 정리 및 label 추가
- portfolioId nullable 처리
- InterviewSessionStatus/InterviewEndType/QuestionCandidateStatus/QuestionCandidateStaleReason에 label 추가
- Answer의 시간 필드(answerStartSec/EndSec/Duration) Integer -> Float
- InterviewAxisPlan: axis -> testType, interviewSessionId -> sessionId
- QuestionCandidate: axis -> testType, secondaryAxis -> secondaryTestType
- Question: indexLevel -> turnLevel, aiVoiceS3Key 추가

+++ src/main/java/com/yapp/d14/interview/domain/Answer.java
@@ -13,9 +13,9 @@ public class Answer {
     private final Long sessionId;
     private final Long questionId;
     private final String sttText;
-    private final Integer answerStartSec;
-    private final Integer answerEndSec;
-    private final Integer answerDuration;
+    private final Float answerStartSec;
+    private final Float answerEndSec;
+    private final Float answerDuration;
     private final Boolean isSkipped;
     private final Float sttFailureRatio;
     private final String evidenceSummary;
@@ -32,9 +32,9 @@ public class Answer {
             Long sessionId,
             Long questionId,
             String sttText,
-            Integer answerStartSec,
-            Integer answerEndSec,
-            Integer answerDuration,
+            Float answerStartSec,
+            Float answerEndSec,
+            Float answerDuration,
             Boolean isSkipped,
             Float sttFailureRatio,
             String evidenceSummary,
@@ -67,9 +67,9 @@ public class Answer {
             Long sessionId,
             Long questionId,
             String sttText,
-            Integer answerStartSec,
-            Integer answerEndSec,
-            Integer answerDuration,
+            Float answerStartSec,
+            Float answerEndSec,
+            Float answerDuration,
             Boolean isSkipped,
             Float sttFailureRatio,
             String evidenceSummary,
@@ -103,9 +103,9 @@ public class Answer {
             Long sessionId,
             Long questionId,
             String sttText,
-            Integer answerStartSec,
-            Integer answerEndSec,
-            Integer answerDuration,
+            Float answerStartSec,
+            Float answerEndSec,
+            Float answerDuration,
             Boolean isSkipped,
             Float sttFailureRatio,
             String evidenceSummary,
